Fix moving AsGatewayConfigurationType attribute to PaymentBundle
