+++ src/Sylius/Bundle/PaymentBundle/DependencyInjection/SyliusPaymentExtension.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\PaymentBundle\DependencyInjection;
 
+use Sylius\Bundle\PaymentBundle\Attribute\AsGatewayConfigurationType;
 use Sylius\Bundle\PaymentBundle\Attribute\AsPaymentMethodsResolver;
 use Sylius\Bundle\ResourceBundle\DependencyInjection\Extension\AbstractResourceExtension;
 use Symfony\Component\Config\FileLocator;
@@ -39,6 +40,17 @@ final class SyliusPaymentExtension extends AbstractResourceExtension
 
     private function registerAutoconfiguration(ContainerBuilder $container): void
     {
+        $container->registerAttributeForAutoconfiguration(
+            AsGatewayConfigurationType::class,
+            static function (ChildDefinition $definition, AsGatewayConfigurationType $attribute): void {
+                $definition->addTag(AsGatewayConfigurationType::SERVICE_TAG, [
+                    'type' => $attribute->getType(),
+                    'label' => $attribute->getLabel(),
+                    'priority' => $attribute->getPriority(),
+                ]);
+            },
+        );
+
         $container->registerAttributeForAutoconfiguration(
             AsPaymentMethodsResolver::class,
             static function (ChildDefinition $definition, AsPaymentMethodsResolver $attribute): void {
