+++ src/Sylius/Bundle/PayumBundle/DependencyInjection/SyliusPayumExtension.php
@@ -15,10 +15,8 @@ namespace Sylius\Bundle\PayumBundle\DependencyInjection;
 
 use Payum\Paypal\ExpressCheckout\Nvp\PaypalExpressCheckoutGatewayFactory;
 use Payum\Stripe\StripeCheckoutGatewayFactory;
-use Sylius\Bundle\PaymentBundle\Attribute\AsGatewayConfigurationType;
 use Sylius\Bundle\ResourceBundle\DependencyInjection\Extension\AbstractResourceExtension;
 use Symfony\Component\Config\FileLocator;
-use Symfony\Component\DependencyInjection\ChildDefinition;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\PrependExtensionInterface;
 use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
@@ -43,27 +41,11 @@ final class SyliusPayumExtension extends AbstractResourceExtension implements Pr
 
         $container->setParameter('payum.template.layout', $config['template']['layout']);
         $container->setParameter('payum.template.obtain_credit_card', $config['template']['obtain_credit_card']);
-
-        $this->registerAutoconfiguration($container);
     }
 
     public function prepend(ContainerBuilder $container): void
     {
         $this->prependSyliusPayment($container);
-    }
-
-    private function registerAutoconfiguration(ContainerBuilder $container): void
-    {
-        $container->registerAttributeForAutoconfiguration(
-            AsGatewayConfigurationType::class,
-            static function (ChildDefinition $definition, AsGatewayConfigurationType $attribute): void {
-                $definition->addTag(AsGatewayConfigurationType::SERVICE_TAG, [
-                    'type' => $attribute->getType(),
-                    'label' => $attribute->getLabel(),
-                    'priority' => $attribute->getPriority(),
-                ]);
-            },
-        );
     }
 
     private function prependSyliusPayment(ContainerBuilder $container): void
